refractored code and added new modus

+++ Arsalan Anwari_SW ARC PS_CynLr SA/types.hpp
@@ -1,14 +1,13 @@
 #pragma once
 
 #include <cinttypes>
-#include <cstddef>
 #include <string_view>
 
 using i8 = std::int8_t;
 using i16 = std::int16_t;
 using i32 = std::int32_t;
 using i64 = std::int64_t;
-using isize = std::ptrdiff_t;
+using isize = std::intmax_t;
 
 using u8 = std::uint8_t;
 using u16 = std::uint16_t;
@@ -25,7 +24,7 @@ using f64 = double;
 using str = std::string;
 using cstr = std::string_view;
 
-using time_ns = i64;
+using time_ns = isize;
